Move CommonStorageResponseHeaders to storage
