+++ sdk/storage/src/data_lake/operations/file_create.rs
@@ -1,8 +1,6 @@
-use azure_core::headers::{
-    etag_from_headers, last_modified_from_headers, CommonStorageResponseHeaders,
-};
-use azure_core::prelude::ContentLength;
-use azure_core::prelude::IfMatchCondition;
+use crate::core::headers::CommonStorageResponseHeaders;
+use azure_core::headers::{etag_from_headers, last_modified_from_headers};
+use azure_core::prelude::*;
 use chrono::{DateTime, Utc};
 use std::convert::TryInto;
