+++ sdk/storage/src/data_lake/responses/set_file_system_properties_response.rs
@@ -1,5 +1,6 @@
+use crate::core::headers::CommonStorageResponseHeaders;
 use azure_core::{
-    headers::{etag_from_headers, last_modified_from_headers, CommonStorageResponseHeaders},
+    headers::{etag_from_headers, last_modified_from_headers},
     Etag,
 };
 use bytes::Bytes;
