+++ sdk/storage/src/table/responses/delete_entity_response.rs
@@ -1,4 +1,4 @@
-use azure_core::headers::CommonStorageResponseHeaders;
+use crate::core::headers::CommonStorageResponseHeaders;
 use bytes::Bytes;
 use http::Response;
 use std::convert::{TryFrom, TryInto};
