+++ sdk/storage_queues/src/responses/create_queue_response.rs
@@ -1,4 +1,4 @@
-use azure_core::headers::CommonStorageResponseHeaders;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use bytes::Bytes;
 use http::response::Response;
 use std::convert::TryInto;
