+++ sdk/storage_queues/src/responses/get_messages_response.rs
@@ -1,5 +1,6 @@
 use crate::PopReceipt;
-use azure_core::headers::{utc_date_from_rfc2822, CommonStorageResponseHeaders};
+use azure_core::headers::utc_date_from_rfc2822;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use azure_storage::core::xml::read_xml;
 use bytes::Bytes;
 use chrono::{DateTime, Utc};
