+++ sdk/storage_queues/src/responses/list_queues_response.rs
@@ -1,5 +1,5 @@
-use azure_core::headers::CommonStorageResponseHeaders;
 use azure_core::prelude::*;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use azure_storage::xml::read_xml;
 use bytes::Bytes;
 use http::response::Response;
