+++ sdk/storage_queues/src/responses/peek_messages_response.rs
@@ -1,4 +1,5 @@
-use azure_core::headers::{utc_date_from_rfc2822, CommonStorageResponseHeaders};
+use azure_core::headers::utc_date_from_rfc2822;
+use azure_storage::core::headers::CommonStorageResponseHeaders;
 use azure_storage::core::xml::read_xml;
 use bytes::Bytes;
 use chrono::{DateTime, Utc};
